Adds links to API keys and tokens documentation
The examples given use a fake API key or a fake token without giving
much information on how to create them. While that information is easy
to find after a quick search, it makes things more convenient to have
links to the documentation readily avalable.

A line about how to get API keys and tokens with links to the Hubspot
website have was added.